// src/app/shared/models/supplier.model.ts
export class SupplierModel {
    SupplierId: number;
    ShortName: string;
    ContactPerson: string;
    address: string;
    city: string;
    reason: string;
    pin: number;
    fax: string;
    gstIn: string;
    regType: string;
    panNum: string;
    tanNum: string;
    state: string;
    dlNum1: string;
    validUpto1: string;
    dlNum2: string;
    validUpto2: string;
    dlNum3: string;
    validUpto3: string;
    dlNum4: string;
    validUpto4: string;
    blocked: string;
    userId: number;
    description:string
    // sCreatedAt: string;
    // sStatus: boolean;
}