// src/app/shared/models/user.model.ts
export class UserModel {
    userId: number;
    roleId: number;
    uniqueId: number;
    fullName: string;
    phone: string;
    mobile: string;
    email: string;
    image: string;
    status: boolean;
    createdAt: string;
    createdBy: number;
    updatedAt: string;
    udatedBy: number;

   
}